Crate pmemblk_sys

FFI bindings to libpmemblk

The libpmemblk library supports arrays of pmem-resident blocks, all the same size, that are atomically updated. For example, a program keeping a cache of fixed-size objects in pmem might find this library useful.

This is not an official port of the NVM Library.

The official libpmemblk documentation can be found at: http://pmem.io/nvml/libpmemblk/

Enums

PMEMblkpool

Statics

PMEMBLK_MIN_POOL

Functions

pmemblk_bsize

pmemblk_check

pmemblk_check_version

pmemblk_close

pmemblk_create

pmemblk_errormsg

pmemblk_nblock

pmemblk_open

pmemblk_read

pmemblk_set_error

pmemblk_set_zero

pmemblk_write